refactor(cortex): consolidate 6-module split into 4 modules
Closes #105. Pre-0.21.0 quality-pass flagged Cortex as living in six
modules with ambiguous boundaries. Consolidated into four with clear
roles:

- daemon/cortex.ts (moved from services/cortex.ts; merges the old
  trigger-cortex-instructions.ts) — orchestration: prompt builder,
  snapshot, task launcher
- context/cortex-brief.ts (absorbed cortex/instructions-input.ts) —
  content assembly: capabilities, delivery decision, retrieval
  guidance, task input prompt
- daemon/api/cortex.ts — HTTP handlers (unchanged role)
- db/queries/cortex-instructions.ts — DB layer (unchanged role)

cortex/ directory removed. Test files moved to match new locations;
no test cases dropped.

Pure refactor; no behavior changes.

Part of v0.22-dev follow-ups.
